Don't drop db in background.
I can't see the point of doing that anymore. It's true that some tests
might be unable to release all connections without much toil, but then
re-creating the db in setUp() would block the start of the next test
anyway, so the only effect of going background is to raise a python
warning.